+++ src/lib/jp2/t1/Codeblock.h
@@ -261,8 +261,8 @@ struct DecompressCodeblock : public Codeblock
 
   private:
 	Segment* segs; /* information on segments */
-	uint32_t numSegments; /* number of segment in block*/
-	uint32_t numSegmentsAllocated; // number of segments allocated for segs array
+	uint8_t numSegments; /* number of segment in block*/
+	uint8_t numSegmentsAllocated; // number of segments allocated for segs array
 };
 
 } // namespace grk
